# golang: ~2x speedup for encodeTensor()