@anight 2017-11-10
@asimshankar Yeah, this is great idea. I'll open a new pr for this one.