@asimshankar 2017-11-14
That test failure seems un-related. Haven't dug in, but it seems that the continuous builds are passing beyond that point, so perhaps something fixed at head.